[MP-49} :recycle: ContentAtachment 별도 엔티티로 분리
