+++ src/main/java/com/pulse/content/domain/Content.java
@@ -21,19 +21,17 @@ public class Content {
     private MemberId memberId;
     private List<Long> hashTagIds;
     private List<ContentCategoryMap> contentCategories;
-    private List<ContentAttachment> contentAttachments;
     private ContentDetail contentDetail;
     private ContentStatus contentStatus;
     private ContentVisibility contentVisibility;
 
     // factory method
-    public static Content of(ContentId contentId, MemberId memberId, List<Long> hashTagIds, List<ContentCategoryMap> contentCategories, List<ContentAttachment> contentAttachments, ContentDetail contentDetail, ContentStatus contentStatus, ContentVisibility contentVisibility) {
+    public static Content of(ContentId contentId, MemberId memberId, List<Long> hashTagIds, List<ContentCategoryMap> contentCategories, ContentDetail contentDetail, ContentStatus contentStatus, ContentVisibility contentVisibility) {
         return Content.builder()
                 .contentId(contentId)
                 .memberId(memberId)
                 .hashTagIds(hashTagIds)
                 .contentCategories(contentCategories)
-                .contentAttachments(contentAttachments)
                 .contentDetail(contentDetail)
                 .contentStatus(contentStatus)
                 .contentVisibility(contentVisibility)
