+++ src/main/java/com/pulse/content/mapper/helper/ContentMapperHelper.java
@@ -1,10 +1,13 @@
 package com.pulse.content.mapper.helper;
 
 import com.pulse.content.adapter.in.web.dto.FileDTO;
+import com.pulse.content.adapter.out.persistence.entity.ContentEntity;
 import com.pulse.content.adapter.out.persistence.entity.vo.ContentAttachmentEntity;
+import com.pulse.content.domain.Content;
 import com.pulse.content.domain.vo.ContentAttachment;
 import com.pulse.content.domain.vo.ContentDetail;
 import com.pulse.content.mapper.ContentAttachmentMapper;
+import com.pulse.content.mapper.ContentMapper;
 import lombok.RequiredArgsConstructor;
 import org.mapstruct.Named;
 import org.springframework.stereotype.Component;
@@ -16,7 +19,20 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class ContentMapperHelper {
 
+    private final ContentMapper contentMapper;
     private final ContentAttachmentMapper contentAttachmentMapper;
+
+
+    /**
+     * 콘텐츠 도메인 객체를 콘텐츠 엔티티 객체로 변환
+     * @param content - 콘텐츠 도메인
+     * @return 콘텐츠 엔티티
+     */
+    @Named("contentDomainToEntity")
+    public ContentEntity contentDomainToEntity(Content content) {
+        return contentMapper.domainToEntity(content);
+    }
+
 
     /**
      * 첨부 파일 도메인 객체를 엔티티 객체로 변환
@@ -28,6 +44,16 @@ public class ContentMapperHelper {
         return contentAttachments.stream()
                 .map(contentAttachmentMapper::domainToEntity)
                 .collect(Collectors.toList());
+    }
+
+    /**
+     * 콘텐츠 엔티티 객체를 콘텐츠 도메인 객체로 변환
+     * @param contentEntity - 콘텐츠 엔티티
+     * @return 콘텐츠 도메인
+     */
+    @Named("contentEntityToDomain")
+    public Content contentEntityToDomain(ContentEntity contentEntity) {
+        return contentMapper.entityToDomain(contentEntity);
     }
 
     /**
